// apps/frontend/common/types.ts
export type CartItemsType = {
  _id: string;
  name: string;
  slug: string;
  image: string;
  price: number;
  salePrice: number;
  isHomeDeliveryFree: boolean;
  status: string;
  quantity: number;
  hasVariant: boolean;
  variant: ProductVariant;
  variant_id: number;
  nameOfVariant: string;
};

export type CartType = {
  user?: string;
  userImage?: string;
  userName: string;
  phoneNumber: string;
  userEmail?: string;

  city: string;
  address: string;
  note: string;

  isPaymentEnabled: boolean;
  paymentMethod: 'cashOnDelivery' | 'onlinePayment';
  isHomeDeliverySelected: boolean;
  isHomeDeliveryFree: boolean;
  total: number;
  shippingCost: number;
  inTotal: number;
  shippingFeeAdvanceForCashOnDelivery: number;
  totalPayableWhileDelivery: number;
  totalExtraDeliveryCharge: number;

  takeShippingFeeAdvance?: boolean;
  cartItems: CartItemsType[];
};

export type OrderType = CartType & {
  _id: string;

  fraudStatus: object;

  status: 'pending' | 'processing' | 'delivered' | 'onDelivery' | 'cancelled';
  paymentStatus: 'pending' | 'success' | 'failed';
  refundStatus: 'processing' | 'failed' | 'success' | '';
  paymentLink: string;
  ssl: {
    bank_tran_id: string;
    refund_ref_id: string;
  };
  steadFast: {
    consignment_id: number;
    tracking_code: string;
    status:
      | 'pending'
      | 'delivered_approval_pending'
      | 'partial_delivered_approval_pending'
      | 'cancelled_approval_pending'
      | 'unknown_approval_pending'
      | 'delivered'
      | 'partial_delivered'
      | 'cancelled'
      | 'hold'
      | 'in_review'
      | 'unknown';
    statusDescription: string;
  };
};